Creating Helper class test. - And Helper class htmlTagGenerator method.
