+++ challenges/Gio-camacho/week-3/pokemon-oop/src/components/details/details.component.js
@@ -1,17 +1,17 @@
-class Details {
-  constructor(tag, parent) {
-    this.tag = tag;
-    this.parent = parent;
-  }
+// class Details {
+//   constructor(tag, parent) {
+//     this.tag = tag;
+//     this.parent = parent;
+//   }
 
-  htmlTagGenerator(innerText, className, href, src) {
-    const createTag = document.createElement(this.tag);
-    createTag.innerText = innerText;
-    createTag.classList = className;
-    createTag.href = href;
-    createTag.src = src;
-    return createTag;
-  }
-}
+//   htmlTagGenerator(innerText, className, href, src) {
+//     const createTag = document.createElement(this.tag);
+//     createTag.innerText = innerText;
+//     createTag.classList = className;
+//     createTag.href = href;
+//     createTag.src = src;
+//     return createTag;
+//   }
+// }
 
-module.exports = Details;
+// module.exports = Details;
