--- challenges/Gio-camacho/week-3/pokemon-oop/src/components/details/details.test.js
@@ -1,20 +1,0 @@
-const Details = require('./details.component');
-
-describe('Given a htmlTagGenerator method', () => {
-  describe('When is invoked', () => {
-    test('Then should return an Html Element', () => {
-      document.body.innerHTML = `
-        <div id="wrapper-details"></div>
-        `;
-
-      const parent = document.getElementById('wrapper-details');
-      const details = new Details('img');
-      const createHtmlTag = details.htmlTagGenerator(null, 'wrapper-details__img', null, './img/picture.png');
-
-      parent.appendChild(createHtmlTag);
-
-      const result = '<img class="wrapper-details__img" src="./img/picture.png">';
-      expect(parent.innerHTML).toContain(result);
-    });
-  });
-});
