+++ challenges/Gio-camacho/week-3/pokemon-oop/src/components/list/list.component.js
@@ -1,0 +1,5 @@
+// const Helper = require('../../services/helper');
+
+const wrapperList = document.querySelector('.wrapper-list');
+
+const helper = new Helper('h1');
