+++ challenges/Gio-camacho/week-3/pokemon-oop/src/components/list/list.html
@@ -9,7 +9,10 @@
 </head>
 
 <body>
-    <div class="wrapper__list"></div>
+    <div class="wrapper-list"></div>
+    <script src="../../services/helper.js"></script>
+    <script src="../../services/services.js"></script>
+    <script src="./list.component.js"></script>
 </body>
 
 </html>
