+++ challenges/Gio-camacho/week-3/pokemon-oop/src/services/helper.js
@@ -1,0 +1,17 @@
+class Helper {
+  constructor(tag, parent) {
+    this.tag = tag;
+    this.parent = parent;
+  }
+
+  htmlTagGenerator(innerText, className, href, src) {
+    const createTag = document.createElement(this.tag);
+    createTag.innerText = innerText;
+    createTag.classList = className;
+    createTag.href = href;
+    createTag.src = src;
+    return createTag;
+  }
+}
+
+module.exports = Helper;
