+++ challenges/Gio-camacho/week-3/pokemon-oop/src/services/services.test.js
@@ -1,4 +1,5 @@
 const Services = require('./services');
+const Details = require('./helper');
 
 describe('Given a getPokemon Object Method', () => {
   describe('When the new Object call the getPokemon() method', () => {
@@ -46,6 +47,25 @@ describe('Given an getAllPokemons Object Method', () => {
       test('Then should return ...', () => {
         expect(allPokemons.count).toBe(1118);
       });
+    });
+  });
+});
+
+describe('Given a htmlTagGenerator method', () => {
+  describe('When is invoked', () => {
+    test('Then should return an Html Element', () => {
+      document.body.innerHTML = `
+        <div id="wrapper-details"></div>
+        `;
+
+      const parent = document.getElementById('wrapper-details');
+      const details = new Details('img');
+      const createHtmlTag = details.htmlTagGenerator(null, 'wrapper-details__img', null, './img/picture.png');
+
+      parent.appendChild(createHtmlTag);
+
+      const result = '<img class="wrapper-details__img" src="./img/picture.png">';
+      expect(parent.innerHTML).toContain(result);
     });
   });
 });
